split edit file
