+++ client/src/Menu.js
@@ -1,11 +1,9 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import { Row, Col } from 'react-grid-system';
-import Holidays from 'date-holidays';
-import CountryCodes from './helper/CountryCodes.js';
-import EditCalendar from './helper/EditCalendar.js';
 import "./css/Menu.css";
-const hd = new Holidays();
+import {Redirect, Link} from 'react-router-dom';
+
 
 class Menu extends Component {
     constructor(props){
@@ -19,7 +17,7 @@ class Menu extends Component {
     onClickEvent = (e) => {
 
       this.setState({menuToggle: 1-this.state.menuToggle});
-      this.props.onClickEventAction(e.target.id);
+      //this.props.onClickEventAction(e.target.id);
    }
 
 
@@ -30,11 +28,11 @@ class Menu extends Component {
             <div className={c}>
                <div className="menu-button" onClick={this.props.onClickToggleMenu}>&lt;</div>
                 <div className="menu-spacer"></div>
-                <a className="menu-topitem" href="#" onClick={this.onClickEvent}><div className="menu-item" id="1">Edit Calendar</div></a>
-                <a className="menu-topitem" href="#" onClick={this.onClickEvent}><div className="menu-item" id="2">Add Event</div></a>
-                <a className="menu-topitem" href="#" onClick={this.onClickEvent}><div className="menu-item" id="3">Add Contributor</div></a>
-                <a className="menu-topitem" href="#" onClick={this.onClickEvent}><div className="menu-item" id="4">Add Holidays</div></a>
-                <a className="menu-topitem" href="#" onClick={this.onClickEvent}><div className="menu-item" id="7">Add New Calendar</div></a>
+                <Link className="menu-topitem" to="/calendar/edit" onClick={this.onClickEvent}><div className="menu-item" id="1">Edit Calendar</div></Link>
+                <Link className="menu-topitem" to="/event/add" onClick={this.onClickEvent}><div className="menu-item" id="2">Add Event</div></Link>
+                <Link className="menu-topitem" to="/calendar/contributor" onClick={this.onClickEvent}><div className="menu-item" id="3">Add Contributor</div></Link>
+                <Link className="menu-topitem" to="/calendar/holidays" onClick={this.onClickEvent}><div className="menu-item" id="4">Add Holidays</div></Link>
+                <Link className="menu-topitem" to="/calendar/add" onClick={this.onClickEvent}><div className="menu-item" id="7">Add New Calendar</div></Link>
                 <div className="menu-spacer"></div>
             </div>
         );
